+++ src/Container/Exception/ContainerException.php
@@ -8,6 +8,7 @@ use Psr\Container\ContainerExceptionInterface;
  * Container Exception
  *
  * A specified exception in handling errors in containers.
+ * NOTE: To be removed in v1.0.0. Use NotFoundException instead.
  *
  * @package Slytherin
  * @author  Rougin Royce Gutib <rougingutib@gmail.com>