+++ src/Container/Exception/NotFoundException.php
@@ -2,16 +2,17 @@
 
 namespace Rougin\Slytherin\Container\Exception;
 
-use Psr\Container\NotFoundExceptionInterface;
+use Rougin\Slytherin\Container\NotFoundException as BaseException;
 
 /**
  * Not Found Exception
  *
  * A specified exception in handling errors in containers.
+ * NOTE: To be removed in v1.0.0. Use Container\NotFoundException instead.
  *
  * @package Slytherin
  * @author  Rougin Royce Gutib <rougingutib@gmail.com>
  */
-class NotFoundException extends \InvalidArgumentException implements NotFoundExceptionInterface
+class NotFoundException extends BaseException
 {
 }