+++ src/Container/NotFoundException.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Rougin\Slytherin\Container;
+
+use Psr\Container\NotFoundExceptionInterface;
+
+/**
+ * Not Found Exception
+ *
+ * A specified exception in handling errors in containers.
+ *
+ * @package Slytherin
+ * @author  Rougin Royce Gutib <rougingutib@gmail.com>
+ */
+class NotFoundException extends \InvalidArgumentException implements NotFoundExceptionInterface
+{
+}